Update README.md
Update readme to align with the new website (https://neuroinformatics.dev/projects/swc-bids.html) and mention the ongoing proposal to extend BIDS to non-human research.

For the most part I have copied verbatim from the website (with a little extra detail on the animal imaging). Maybe this is not a good idea, on the plus side it is nice to have a single description of the project we can use. On the negative side it might look a bit strange if you navigate from the website to the github.